#51: Checking if image is an illustration in separate process to prevent imagick memory footprint grow
