+++ Dockerfile
@@ -109,6 +109,10 @@ RUN mkdir -p /go/src/github.com/Pixboost/
 WORKDIR /go/src/github.com/Pixboost/
 RUN git clone https://github.com/Pixboost/transformimgs.git
 
+WORKDIR /go/src/github.com/Pixboost/transformimgs/illustration
+
+RUN go install
+
 WORKDIR /go/src/github.com/Pixboost/transformimgs/cmd
 
 RUN go build -o /transformimgs
