+++ Dockerfile.dev
@@ -105,8 +105,15 @@ ENV PATH $GOPATH/bin:$PATH
 RUN mkdir -p "$GOPATH/src" "$GOPATH/bin" && chmod -R 1777 "$GOPATH"
 WORKDIR $GOPATH
 
+
+
 ENV IM_HOME /usr/local/bin
 
 VOLUME /go/src/github.com/Pixboost/transformimgs/
 WORKDIR /go/src/github.com/Pixboost/transformimgs/
+
+RUN cd illustration && \
+    go install && \
+    cd ../
+
 ENTRYPOINT ["sh", "/go/src/github.com/Pixboost/transformimgs/run.sh"]
