+++ illustration/go.mod
@@ -1,0 +1,5 @@
+module github.com/Pixboost/illustration
+
+go 1.18
+
+require gopkg.in/gographics/imagick.v3 v3.7.0 // indirect
